Add edit playlist title
